+++ python-libraries/data-platform-catalogue/data_platform_catalogue/client/datahub/search.py
@@ -2,7 +2,7 @@ import json
 import logging
 from datetime import datetime
 from importlib.resources import files
-from typing import Any, Sequence
+from typing import Any, Sequence, Tuple
 
 from datahub.configuration.common import GraphError
 from datahub.ingestion.graph.client import DataHubGraph
@@ -184,14 +184,19 @@ class SearchClient:
                 tags.append(properties["name"])
         return tags
 
-    def _parse_properties(self, entity: dict[str, Any]) -> dict[str, Any]:
+    def _parse_properties(
+        self, entity: dict[str, Any]
+    ) -> Tuple[dict[str, Any], dict[str, Any]]:
         """
         Parse properties and editableProperties into a single dictionary.
         """
         properties = entity["properties"] or {}
         editable_properties = entity.get("editableProperties") or {}
         properties.update(editable_properties)
-        return properties
+        custom_properties = {
+            i["key"]: i["value"] for i in properties.get("customProperties", [])
+        }
+        return properties, custom_properties
 
     def _parse_domain(self, entity: dict[str, Any]):
         metadata = {}
@@ -199,9 +204,8 @@ class SearchClient:
         inner_domain = domain.get("domain") or {}
         metadata["domain_id"] = inner_domain.get("urn", "")
         if inner_domain:
-            metadata["domain_name"] = self._parse_properties(inner_domain).get(
-                "name", ""
-            )
+            domain_properties, _ = self._parse_properties(inner_domain)
+            metadata["domain_name"] = domain_properties.get("name", "")
         else:
             metadata["domain_name"] = ""
         return metadata
@@ -211,7 +215,7 @@ class SearchClient:
         Map a dataset entity to a SearchResult
         """
         owner_email, owner_name = self._parse_owner(entity)
-        properties = self._parse_properties(entity)
+        properties, custom_properties = self._parse_properties(entity)
         tags = self._parse_tags(entity)
         last_updated = self._parse_last_updated(entity)
         name = entity["name"]
@@ -230,6 +234,7 @@ class SearchClient:
             "data_products": data_products,
         }
         metadata.update(self._parse_domain(entity))
+        metadata.update(custom_properties)
 
         return SearchResult(
             id=entity["urn"],
@@ -247,7 +252,7 @@ class SearchClient:
         Map a data product entity to a SearchResult
         """
         owner_email, owner_name = self._parse_owner(entity)
-        properties = self._parse_properties(entity)
+        properties, custom_properties = self._parse_properties(entity)
         tags = self._parse_tags(entity)
         last_updated = self._parse_last_updated(entity)
         metadata = {
@@ -256,6 +261,7 @@ class SearchClient:
             "number_of_assets": properties["numAssets"],
         }
         metadata.update(self._parse_domain(entity))
+        metadata.update(custom_properties)
 
         return SearchResult(
             id=entity["urn"],